pyproject.toml
[tool.poetry]
name = "citoplasm"
version = "0.0.5"
description = "CITOplasm is a Python library for writing LLM code in a declarative way."
authors = ["Josh Pollock <jopo@mit.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.12"
anyio = "3.6.2"
attrs = "22.2.0"
black = "23.3.0"
certifi = "2022.12.7"
charset-normalizer = "3.1.0"
click = "8.1.3"
contourpy = "1.0.7"
cycler = "0.11.0"
defopt = "6.4.0"
docutils = "0.19"
exceptiongroup = "1.1.1"
executing = "1.2.0"
Faker = "18.3.2"
fastapi = "0.95.0"
filelock = "3.10.7"
fonttools = "4.39.3"
fvalues = "0.0.3"
h11 = "0.14.0"
httpcore = "0.16.3"
httpx = "0.23.3"
huggingface-hub = "0.13.3"
idna = "3.4"
iniconfig = "2.0.0"
joblib = "1.2.0"
kiwisolver = "1.4.4"
Levenshtein = "0.20.9"
markdown-it-py = "2.2.0"
matplotlib = "3.7.1"
mdurl = "0.1.2"
merge-args = "0.1.5"
more-itertools = "9.1.0"
mypy-extensions = "1.0.0"
nest-asyncio = "1.5.6"
nltk = "3.8.1"
numerizer = "0.2.2"
numpy = "1.24.2"
ought-ice = "0.5.0"
packaging = "23.0"
pandas = "2.0.0"
pathspec = "0.11.1"
Pillow = "9.5.0"
platformdirs = "3.2.0"
pluggy = "1.0.0"
pockets = "0.9.1"
prompt-toolkit = "3.0.38"
pydantic = "1.10.7"
Pygments = "2.14.0"
pyparsing = "3.0.9"
pytest = "7.2.2"
pytest-asyncio = "0.21.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-Levenshtein = "0.20.9"
python-ulid = "1.1.0"
pytz = "2023.3"
PyYAML = "6.0"
questionary = "1.10.0"
rapidfuzz = "2.15.0"
regex = "2023.3.23"
rfc3986 = "1.5.0"
rich = "13.3.3"
rouge-metric = "1.0.1"
scikit-learn = "1.2.2"
scipy = "1.10.1"
six = "1.16.0"
sniffio = "1.3.0"
sphinxcontrib-napoleon = "0.7"
starlette = "0.26.1"
structlog = "22.3.0"
tenacity = "8.2.2"
thefuzz = "0.19.0"
threadpoolctl = "3.1.0"
tokenizers = "0.13.2"
tomli = "2.0.1"
tqdm = "4.65.0"
transformers = "4.27.4"
typing_extensions = "4.5.0"
tzdata = "2023.3"
urllib3 = "1.26.15"
uvicorn = "0.21.1"
wcwidth = "0.2.6"
websockets = "11.0"

[tool.poetry.group.test.dependencies]
requests = "^2.29.0"
mwparserfromhell = "^0.6.4"
ipykernel = "^6.22.0"
ipywidgets = "^8.0.6"
types-requests = "^2.29.0.0"
beautifulsoup4 = "^4.12.2"
streamlit = "^1.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
